Rename utils.js utils.ts
